+++ vt_graph_api/load/maltego/csv.py
@@ -1,9 +1,9 @@
-"""vt_graph_api.load.maltego_csv.
+"""vt_graph_api.load.maltego.csv.
 
 This modules provides graph loader method for
 maltego graph in csv format.
 """
 
 
-def from_maltego_csv():
+def from_csv():
   pass