+++ vt_graph_api/load/virustotal/__init__.py
@@ -1,0 +1,10 @@
+"""vt_graph_api.load.virustotal.
+
+This modules provides virustotal graph loaders for virutstotal graph api.
+"""
+
+
+from vt_graph_api.load.virustotal.graph_id import from_graph_id
+
+
+__all__ = ["from_graph_id"]