+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 - Added regular expression to detect node types.
 - Accept **MD5** and **SHA1** as valid ID for nodes with **file type**.
 - Added **VTIntelligence** search for nodes without any information.
+- Accept custom node types.
 
 #### Load
 ---
-- Added load graph from **VTGraph** 
+- Added load graph from **VTGraph**.
+- Added load graph from **Maltego graphml**.